feat: refactor store structure by moving to a centralized store and adding user and contract slices
